feat(playlist-maker): move combobox items fetch into hooks
