+++ custom-apps/playlist-maker/src/components/nodes/sources/RecommendedPlaylistTracksSourceNode.tsx
@@ -1,7 +1,9 @@
-import { searchDesktop } from '@shared/graphQL/queries/search-desktop';
-import { getPlatform } from '@shared/utils/spicetify-utils';
 import { useComboboxValues } from 'custom-apps/playlist-maker/src/hooks/use-combobox-values';
 import { useNodeForm } from 'custom-apps/playlist-maker/src/hooks/use-node-form';
+import {
+    type PlaylistItem,
+    usePlaylistComboboxFetchers,
+} from 'custom-apps/playlist-maker/src/hooks/use-playlist-combobox-fetchers';
 import {
     RecommendedPlaylistTracksDataSchema,
     type RecommendedPlaylistTracksData,
@@ -20,28 +22,19 @@ import { NodeField } from '../shared/NodeField';
 import { SourceNodeHeader } from '../shared/NodeHeader';
 import { NodeTitle } from '../shared/NodeTitle';
 
-type PlaylistItem = {
-    id: string;
-    uri: string;
-    name: string;
-    image: string | null;
-    ownerName: string;
-};
-
 function PlaylistItemRenderer(
     props: Readonly<ItemRendererProps<PlaylistItem>>,
 ): JSX.Element {
     return (
         <div className="flex max-h-[80px] items-stretch gap-2">
             <div className="flex h-[60px] w-[60px] shrink-0 items-center justify-center !p-2">
-                {props.item.image && (
+                {props.item.image ? (
                     <img
                         src={props.item.image}
                         className="rounded-md object-contain"
                         alt="playlist"
                     />
-                )}
-                {props.item.image === null && (
+                ) : (
                     <Music size={60} strokeWidth={1} />
                 )}
             </div>
@@ -76,76 +69,12 @@ export function RecommendedPlaylistTracksSourceNode(
             RecommendedPlaylistTracksDataSchema,
         );
 
-    const getPlaylists = useCallback(
-        async (input: string): Promise<PlaylistItem[]> => {
-            if (!input.trim()) {
-                return [];
-            }
+    const onPlaylistNotFound = useCallback(() => {
+        updateNodeField({ playlistUri: '' });
+    }, [updateNodeField]);
 
-            const search = await searchDesktop({
-                searchTerm: input,
-                offset: 0,
-                limit: 20,
-                includePreReleases: true,
-                includeArtistHasConcertsField: false,
-                includeAudiobooks: false,
-                includeLocalConcertsField: false,
-                numberOfTopResults: 5,
-            });
-
-            const items: PlaylistItem[] = search.searchV2.playlists.items.map(
-                (playlist) => ({
-                    id: playlist.data.uri,
-                    uri: playlist.data.uri,
-                    name: playlist.data.name,
-                    image:
-                        playlist.data.images.items[0]?.sources[0].url ?? null,
-                    ownerName: playlist.data.ownerV2.data.name,
-                }),
-            );
-
-            return items;
-        },
-        [],
-    );
-
-    const getPlaylist = useCallback(
-        async (playlistUri: string): Promise<PlaylistItem | null> => {
-            const playlistApi = getPlatform().PlaylistAPI;
-
-            try {
-                const playlist = await playlistApi.getPlaylist(
-                    playlistUri,
-                    {},
-                    {},
-                );
-
-                const playlistItem: PlaylistItem = {
-                    id: playlist.metadata.uri,
-                    name: playlist.metadata.name,
-                    uri: playlist.metadata.uri,
-                    image:
-                        playlist.metadata.images.length > 0
-                            ? playlist.metadata.images[0].url
-                            : null,
-                    ownerName: playlist.metadata.owner.displayName,
-                };
-
-                return playlistItem;
-            } catch (e) {
-                console.error('Failed to fetch playlist', e);
-                updateNodeField({ playlistUri: '' });
-
-                return null;
-            }
-        },
-        [updateNodeField],
-    );
-
-    const itemToString = useCallback(
-        (item: PlaylistItem): string => item.name,
-        [],
-    );
+    const { getPlaylist, getPlaylists, itemToString } =
+        usePlaylistComboboxFetchers(onPlaylistNotFound);
 
     const {
         inputValue,
