+++ custom-apps/playlist-maker/src/components/nodes/sources/SearchAlbumSourceNode.tsx
@@ -1,9 +1,7 @@
-import { getAlbum as getGraphQlAlbum } from '@shared/graphQL/queries/get-album';
-import { searchDesktop } from '@shared/graphQL/queries/search-desktop';
 import {
-    type AlbumResponseWrapper,
-    type PreReleaseResponseWrapper,
-} from '@shared/graphQL/types/search-desktop-data-albums';
+    type AlbumItem,
+    useAlbumComboboxFetchers,
+} from 'custom-apps/playlist-maker/src/hooks/use-album-combobox-fetchers';
 import { useComboboxValues } from 'custom-apps/playlist-maker/src/hooks/use-combobox-values';
 import { useNodeForm } from 'custom-apps/playlist-maker/src/hooks/use-node-form';
 import {
@@ -26,28 +24,19 @@ import { NodeField } from '../shared/NodeField';
 import { SourceNodeHeader } from '../shared/NodeHeader';
 import { NodeTitle } from '../shared/NodeTitle';
 
-type AlbumItem = {
-    id: string;
-    uri: string;
-    name: string;
-    image: string | null;
-    artists: string;
-};
-
 function AlbumItemRenderer(
     props: Readonly<ItemRendererProps<AlbumItem>>,
 ): JSX.Element {
     return (
         <div className="flex max-h-[80px] items-stretch gap-2">
             <div className="flex h-[60px] w-[60px] shrink-0 items-center justify-center !p-2">
-                {props.item.image && (
+                {props.item.image ? (
                     <img
                         src={props.item.image}
                         className="rounded-md object-contain"
                         alt="album"
                     />
-                )}
-                {props.item.image === null && (
+                ) : (
                     <Music size={60} strokeWidth={1} />
                 )}
             </div>
@@ -78,110 +67,12 @@ export function SearchAlbumSourceNode(
         AlbumDataSchema,
     );
 
-    const getAlbums = useCallback(
-        async (input: string): Promise<AlbumItem[]> => {
-            if (!input.trim()) {
-                return [];
-            }
+    const onAlbumNotFound = useCallback(() => {
+        updateNodeField({ uri: '' });
+    }, [updateNodeField]);
 
-            const search = await searchDesktop({
-                searchTerm: input,
-                offset: 0,
-                limit: 10,
-                includePreReleases: true,
-                includeArtistHasConcertsField: false,
-                includeAudiobooks: false,
-                includeLocalConcertsField: false,
-                numberOfTopResults: 5,
-            });
-
-            const items: AlbumItem[] = search.searchV2.albumsV2.items.map(
-                (album) => {
-                    const albumIsPrelease = (
-                        album: PreReleaseResponseWrapper | AlbumResponseWrapper,
-                    ): album is PreReleaseResponseWrapper => {
-                        return album.__typename === 'PreReleaseResponseWrapper';
-                    };
-
-                    if (albumIsPrelease(album)) {
-                        return {
-                            id: album.data.uri,
-                            uri: album.data.uri,
-                            name: album.data.preReleaseContent.name,
-                            image:
-                                album.data.preReleaseContent.coverArt.sources
-                                    .length > 0
-                                    ? album.data.preReleaseContent.coverArt
-                                          .sources[0].url
-                                    : null,
-                            artists: album.data.preReleaseContent.artists.items
-                                .map((artist) => artist.data.profile.name)
-                                .join(', '),
-                        };
-                    } else {
-                        return {
-                            id: album.data.uri,
-                            uri: album.data.uri,
-                            name: album.data.name,
-                            image:
-                                album.data.coverArt.sources.length > 0
-                                    ? album.data.coverArt.sources[0].url
-                                    : null,
-                            artists: album.data.artists.items
-                                .map((artist) => artist.profile.name)
-                                .join(', '),
-                        };
-                    }
-                },
-            );
-
-            return items;
-        },
-        [],
-    );
-
-    const getAlbum = useCallback(
-        async (albumUri: string): Promise<AlbumItem | null> => {
-            try {
-                const album = await getGraphQlAlbum({
-                    uri: albumUri,
-                    offset: 0,
-                    limit: 0,
-                    locale: Spicetify.Locale.getLocale(),
-                });
-
-                if (album.albumUnion.__typename === 'NotFound') {
-                    throw new Error('Album not found');
-                }
-
-                const albumItem: AlbumItem = {
-                    id: album.albumUnion.uri,
-                    name: album.albumUnion.name,
-                    uri: album.albumUnion.uri,
-                    image:
-                        album.albumUnion.coverArt.sources.length > 0
-                            ? album.albumUnion.coverArt.sources[0].url
-                            : null,
-                    artists: album.albumUnion.artists.items
-                        .map((artist) => artist.profile.name)
-                        .join(', '),
-                };
-
-                return albumItem;
-            } catch (e) {
-                console.error('Failed to fetch album', e);
-                updateNodeField({ uri: '' });
-
-                return null;
-            }
-        },
-        [updateNodeField],
-    );
-
-    const itemToString = useCallback(
-        (item: AlbumItem): string => item.name,
-        [],
-    );
+    const { getAlbum, getAlbums, itemToString } =
+        useAlbumComboboxFetchers(onAlbumNotFound);
 
     const {
         inputValue,
